invalid read solved
