--- sources/tokenizer/split_tokens.c
@@ -1,51 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   split_tokens.c                                     :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: phemsi-a <phemsi-a@student.42sp.org.br>    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/06/22 18:34:41 by phemsi-a          #+#    #+#             */
-/*   Updated: 2021/07/05 08:29:20 by phemsi-a         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minishell.h"
-
-static int	find_end(char *line, int i, int *end)
-{
-	while (!(ft_isblank(line[i])) && (line[i]))
-	{
-		if (is_between_quotes(line, i, &i, line[i]))
-		{
-			i++;
-			continue ;
-		}
-		i++;
-	}
-	*end = i;
-	return (i);
-}
-
-t_token	*split_tokens(char *line)
-{
-	int		i;
-	int		token_end;
-	t_token	*token_lst;
-
-	i = 0;
-	token_end = i;
-	token_lst = NULL;
-	while (line[i])
-	{
-		while (ft_isblank(line[i]))
-			i++;
-		if (!line[i])
-			break ;
-		add_token(line, i, find_end(line, i, &token_end), &token_lst);
-		i = token_end;
-		if (line[i])
-			i++;
-	}
-	return (token_lst);
-}
